fix: Memory element type
change i32 to i64

+++ 2019/day02/src/lib.rs
@@ -1,12 +1,12 @@
 use std::path::PathBuf;
 use common::computer::{computer, read_program_file};
 
-pub fn part_1(input_1: u32, input_2: u32) -> Vec<i32> {
+pub fn part_1(input_1: u32, input_2: u32) -> Vec<i64> {
     let path = PathBuf::from("./assets/program.txt");
     let mut memory = read_program_file(path).unwrap();
     // Define inputs
-    memory[1] = input_1 as i32;
-    memory[2] = input_2 as i32;
+    memory[1] = input_1 as i64;
+    memory[2] = input_2 as i64;
     // Compute
     let (memory, _) = computer(memory, None);
     memory
@@ -20,10 +20,10 @@ pub fn part_2(result: u32) -> Option<(u32, u32)> {
     for i in 0..100 as u32 {
         for j in 0..100 as u32 {
             let mut memory = memory_template.clone();
-            memory[1] = i as i32;
-            memory[2] = j as i32;
+            memory[1] = i as i64;
+            memory[2] = j as i64;
             let (memory, _) = computer(memory, None);
-            if memory[0] == result as i32{
+            if memory[0] == result as i64{
                 return Some((i, j))
             }
         }
