+++ 2019/day05/src/lib.rs
@@ -2,7 +2,7 @@ use std::path::PathBuf;
 use common::computer::{read_program_file, computer};
 use std::collections::VecDeque;
 
-pub fn part_1() -> Vec<i32> {
+pub fn part_1() -> Vec<i64> {
     let path = PathBuf::from("./assets/program.txt");
     let memory = read_program_file(path).unwrap();
 
@@ -10,7 +10,7 @@ pub fn part_1() -> Vec<i32> {
     outputs
 }
 
-pub fn part_2() -> Vec<i32> {
+pub fn part_2() -> Vec<i64> {
     let path = PathBuf::from("./assets/program.txt");
     let memory = read_program_file(path).unwrap();
